requirements.txt
numpy>=1.24
scipy>=1.10
h5py>=3.8
Pillow>=9.5
tqdm>=4.65
scikit-learn>=1.3
medpy>=0.5.1
segmentation-models-pytorch>=0.3.3
timm>=0.9.12
torch>=2.1
torchvision>=0.16